refactor: move setupkeys/sshconfig -> ssh/sshconfig
