+++ cmd/topo/setup_ssh.go
@@ -3,8 +3,8 @@ package main
 import (
 	"os"
 
-	"github.com/arm/topo/internal/setupkeys/sshconfig"
 	"github.com/arm/topo/internal/ssh"
+	"github.com/arm/topo/internal/ssh/sshconfig"
 	"github.com/spf13/cobra"
 )
